Renamed `runtime` package to `stub` and replaced all corresponding imports
