+++ core/stub/config/logger.go
