+++ core/stub/range_getter.go
@@ -1,4 +1,4 @@
-package golang
+package stub
 
 import (
 	"fmt"
