+++ core/stub/range_getter_test.go
@@ -1,4 +1,4 @@
-package golang
+package stub
 
 import (
 	"testing"
